Change how reciprocal is computed
I have not fully understood why this solve the problem, but splitting
the operation in 2 lines does not seem to work

+++ feature_engine/transformation/reciprocal.py
@@ -152,8 +152,7 @@ class ReciprocalTransformer(BaseNumericalTransformer):
 
         # transform
         # for some reason reciprocal does not work with integers
-        X.loc[:, self.variables_] = X.loc[:, self.variables_].astype("float")
-        X.loc[:, self.variables_] = np.reciprocal(X.loc[:, self.variables_])
+        X.loc[:, self.variables_] = np.reciprocal(X.loc[:, self.variables_].astype("float"))
 
         return X
